CI: Test multizone interconnect.
3 existing shard-confermance jobs and 3 existing control-plane jobs
are now testd with single node zone interconnect deployment.

Also adds a new e2e control plane job 'multi-node-zones' which deploys
a 2 zone and 2 node per zone deployment (ovn-control-plane and
ovn-worker in zone - "zone-1" and ovn-worker2 and ovn-worker3 nodes
in the zone - "zone-2") and runs a simple test. For now a single test
is added which creates a server pod in one zone and a client pod in
other zone and
  - tests connectivity between the two
  - moves the client pod scheduled node to the server pod schedule node
    zone and tests the connectivity
  - moves back the client pod node to its original zone and tests the
    connectivity again.

[Note: This patch also skips e2e control plante tests for egress ip, egress
services and multihoming layer2 and localet networks for multi zone
setup. This is for testing only. Once all the tests pass I'll rebase
this PR on top of the egress-ip multizone PR -
https://github.com/ovn-org/ovn-kubernetes/pull/3561]

+++ test/e2e/egress_services.go
@@ -18,6 +18,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/kubernetes/test/e2e/framework"
 	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
+	e2eskipper "k8s.io/kubernetes/test/e2e/framework/skipper"
 	utilnet "k8s.io/utils/net"
 )
 
@@ -38,6 +39,7 @@ var _ = ginkgo.Describe("Egress Services", func() {
 		externalIPv4 string
 		externalIPv6 string
 		nodes        []v1.Node
+		testsSkipped bool
 	)
 
 	f := wrappedTestFramework("egress-services")
@@ -52,6 +54,17 @@ var _ = ginkgo.Describe("Egress Services", func() {
 				"Test requires >= 3 Ready nodes, but there are only %v nodes",
 				len(n.Items))
 		}
+		testsSkipped = true
+		multiZones, err := isMultipleZoneDeployment(f.ClientSet)
+		if err != nil {
+			framework.Failf("Failed to get the node zones : %v", err)
+		}
+		if multiZones {
+			e2eskipper.Skipf(
+				"Egress services are not yet supported with multiple zones deployment",
+			)
+		}
+		testsSkipped = false
 		nodes = n.Items
 		ginkgo.By("Creating an external container to send the traffic to/from")
 		externalIPv4, externalIPv6 = createClusterExternalContainer(externalContainerName, agnhostImage,
@@ -60,8 +73,10 @@ var _ = ginkgo.Describe("Egress Services", func() {
 	})
 
 	ginkgo.AfterEach(func() {
-		deleteClusterExternalContainer(externalContainerName)
-		flushCustomRoutingTableOnNodes(nodes, customRoutingTable)
+		if !testsSkipped {
+			deleteClusterExternalContainer(externalContainerName)
+			flushCustomRoutingTableOnNodes(nodes, customRoutingTable)
+		}
 	})
 
 	ginkgotable.DescribeTable("Should validate pods' egress is SNATed to the LB's ingress ip without selectors",
